+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,26 @@
 # ai
 
+## 5.0.0-beta.6
+
+### Patch Changes
+
+- 143c55b: feat (ai): export Chat callback types
+- f04ffe4: feat (ui): add onData callback to Chat
+- 97c35c0: feat (ui): transient data parts
+- fccf75c: update mcp protocol version
+
+## 5.0.0-beta.5
+
+### Patch Changes
+
+- 4f3e637: fix (ui): avoid caching globalThis.fetch in case it is patched by other libraries
+
+## 5.0.0-beta.4
+
+### Patch Changes
+
+- 09f41ac: fix (ui): add message metadata in Chat.sendMessage
+
 ## 5.0.0-beta.3
 
 ### Patch Changes